refactor(stdlib): split more builtins into own files
